% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllDataFromServer.R
\name{fetchAllDataFromServer}
\alias{fetchAllDataFromServer}
\title{Creates the cohorts needed for the abalysis and the gets the patient level prediction data from the server}
\usage{
fetchAllDataFromServer(connectionDetails = connectionDetails,
  cdmDatabaseSchema = cdmDatabaseSchema,
  oracleTempSchema = oracleTempSchema,
  workDatabaseSchema = workDatabaseSchema,
  studyCohortTable = studyCohortTable, workFolder = workFolder,
  verbosity = INFO)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{workDatabaseSchema}{The name of the database schema that is the location where the
cohort data used to define the study cohorts is available}

\item{studyCohortTable}{The tablename that contains the study cohorts.}

\item{workFolder}{The directory where the results will be saved to}
}
\value{
Returns TRUE when finished
}
\description{
This function creates the 'at risk' and 'outcome' cohorts using files saved in the package to the
 workDatabaseSchema database in the studyCohortTable table and then extracts the plpData
 using these cohorts.
}
\details{
This function creates the 'at risk' and 'outcome' cohorts using files saved in the package to the
 workDatabaseSchema database in the studyCohortTable table and then extracts the plpData
 using these cohorts.
}
