% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRFPredictionModels.R
\name{fitRFPredictionModels}
\alias{fitRFPredictionModels}
\title{Creates the random forest models for all outcomes}
\usage{
fitRFPredictionModels(workFolder)
}
\arguments{
\item{workFolder}{The directory where the plpData and population are saved to}
}
\value{
Returns TRUE when finished and saves the models  into the workFolder directory
in the subdirectory named rfModels
}
\description{
This function creates the random forest models for all outcomes
}
\details{
The patient level prediction runPlp function is called for each outcome to train
a random forest model
}
