% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAllPopulations.R
\name{generateAllPopulations}
\alias{generateAllPopulations}
\title{Creates the populations for the predictions for each outcome}
\usage{
generateAllPopulations(workFolder, verbosity = INFO)
}
\arguments{
\item{workFolder}{The directory where the plpData is saved to}
}
\value{
Returns TRUE when finished and saves the populations as csvs into the workFolder directory
in the subdirectory names populations
}
\description{
This function creates the popualtions for each outcome in the plpData
}
\details{
For each outcome the population of people who fit the prediction criteria are found
and the people who have the outcome during the year after the 'at risk' cohort are
identified
}
