% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitResults.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(workFolder, dbName)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes subfolder export will be created there
(/)}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
